@@ -14,16 +14,14 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  arcore_flutter_plugin:
+  vector_math: ^2.1.0
+  cupertino_icons: ^1.0.8
+  arcore_flutter_plugin: 0.1.0
 
 dev_dependencies:
   flutter_test:
