@@ -18,7 +18,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
 
   @override
@@ -26,66 +26,54 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  ArCoreController arCoreController;
+  late ArCoreController arCoreController;
 
+  @override
+  void dispose() {
+    // Siin kontrollime enne, kas arCoreController on olemas
+    arCoreController.dispose();
+    super.dispose();
+  }
+
+  // AR vaate loomise meetod
   _onArCoreViewCreated(ArCoreController _arcoreController) {
     arCoreController = _arcoreController;
     _addSphere(arCoreController);
     _addCube(arCoreController);
-    _addCyclinder(arCoreController);
+    _addCylinder(arCoreController);
   }
 
+  // Sfääri lisamine AR vaatesse
   _addSphere(ArCoreController _arcoreController) {
     final material = ArCoreMaterial(color: Colors.deepPurple);
     final sphere = ArCoreSphere(materials: [material], radius: 0.2);
     final node = ArCoreNode(
       shape: sphere,
-      position: vector.Vector3(
-        0,
-        0,
-        -1,
-      ),
+      position: vector.Vector3(0, 0, -1),
     );
-
-    _arcoreController.add(node);
+    _arcoreController.addArCoreNode(node);
   }
 
-  _addCyclinder(ArCoreController _arcoreController) {
+  // Silindri lisamine AR vaatesse
+  _addCylinder(ArCoreController _arcoreController) {
     final material = ArCoreMaterial(color: Colors.green, reflectance: 1);
-    final cylinder =
-        ArCoreCylinder(materials: [material], radius: 0.4, height: 0.3);
+    final cylinder = ArCoreCylinder(materials: [material], radius: 0.4, height: 0.3);
     final node = ArCoreNode(
       shape: cylinder,
-      position: vector.Vector3(
-        0,
-        -2.5,
-        -3.0,
-      ),
+      position: vector.Vector3(0, -2.5, -3.0),
     );
-
-    _arcoreController.add(node);
+    _arcoreController.addArCoreNode(node);
   }
 
+  // Kubi lisamine AR vaatesse
   _addCube(ArCoreController _arcoreController) {
     final material = ArCoreMaterial(color: Colors.pink, metallic: 1);
-    final cube =
-        ArCoreCube(materials: [material], size: vector.Vector3(1, 1, 1));
+    final cube = ArCoreCube(materials: [material], size: vector.Vector3(1, 1, 1));
     final node = ArCoreNode(
       shape: cube,
-      position: vector.Vector3(
-        -0.5,
-        -0.5,
-        -3,
-      ),
+      position: vector.Vector3(-0.5, -0.5, -3),
     );
-
-    _arcoreController.add(node);
-  }
-
-  @override
-  void dispose() {
-    arCoreController.dispose();
-    super.dispose();
+    _arcoreController.addArCoreNode(node);
   }
 
   @override
